% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPrimer.R
\name{checkPrimer}
\alias{checkPrimer}
\title{checkPrimer}
\usage{
checkPrimer(fwdPrimer, revPrimer, exp, ITR = "PiggyBac")
}
\arguments{
\item{fwdPrimer}{A character-string of the forward-primer used.}

\item{revPrimer}{A character-string of the reverse-primer used.}

\item{exp}{The tagMeppr-object of a sample: first run \code{\link{align}}.}

\item{ITR}{Can take PiggyBac (default), SleepingBeauty, or a path to a 1000xN-padded ITR.fasta.}
}
\value{
The experiment-object will be updated with the rev5_fwd3-flag, which
will tell all downstream analyses if our assumptions are correct.
}
\description{
This tool checks the assumptions about the primers.
}
\details{
The expected general layout for the ITR-sequence looks like this:

\code{|---ITR---NNN...NNN---ITR---|}

The primers are expected to be 5'-end for the reverse and 3' for the forward:

\code{<rev}
\code{|---ITR---NNN...NNN---ITR---|}
\code{                         fwd>}

This tool checks these assumptions and sets the rev5_fwd3 flag to TRUE.
}
\examples{
\dontrun{

C9 = newTagMeppr(F1 = 'clone9_FWD_R1.fq.gz',
                 F2 = 'clone9_FWD_R2.fq.gz',
                 R1 = 'clone9_REV_R1.fq.gz',
                 R2 = 'clone9_REV_R2.fq.gz',
                 name = "clone9",
                 protocol = 'PiggyBac')

checkPrimer(fwdPrimer = "CGTCAATTTTACGCAGACTATC",
            revPrimer = "GTACGTCACAATATGATTATCTTTCTAG",
            exp =  C9,
            ITR = 'PiggyBac')

}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
