% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInsertions.R
\name{findInsertions}
\alias{findInsertions}
\title{findInsertions}
\usage{
findInsertions(exp, ref, padding = NULL)
}
\arguments{
\item{exp}{The tagMeppr-object of a sample: first run \code{\link{align}}.}

\item{ref}{A reference object of \code{\link{makeIndex}} or \code{\link{loadIndex}}.}

\item{padding}{Add padding to the start/end of reads in relation to the TIS.
Default is half of the TIS (i.e. 2bp whit Piggybac), meaning that the mapper
can report a match Xbp further than the TIS (i.e. TTAA).}
}
\value{
The object will be updated with a results-table:
\describe{
\item{fwdCount}{Forward read-counts at this insertion-site}
\item{fwdOrient}{Forward read-orientation at the insertion}
\item{D_FWD}{The imbalance of forward reads on both sides of the
site: D = 0 means no imbalance}
\item{fwdBinom}{The probability of the imbalance}
\item{revCount}{Reverse read-counts at this insertion-site}
\item{revOrient}{Reverse read-orientation at the insertion}
\item{D_REV}{The imbalance of forward reads on both sides of the site: D = 0
means no imbalance}
\item{revBinom}{The probability of the imbalance}
\item{pMarkerMeppR}{The combined probability of the imbalances}
\item{padj}{Bonferroni-adjusted probabilties}
\item{orientation}{The orientation of the insertion}
}
}
\description{
Identify highly likely insertion-sites in an tagMeppr-object.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C9 = newTagMeppr(F1 = 'clone9_FWD_R1.fq.gz',
                 F2 = 'clone9_FWD_R2.fq.gz',
                 R1 = 'clone9_REV_R1.fq.gz',
                 R2 = 'clone9_REV_R2.fq.gz',
                 name = "clone9",
                 protocol = 'PiggyBac')

align(exp = C9, ref = reference_hg19_PB, cores = 30, empericalCentre = T)

findInsertions(exp = C9, ref = reference_hg19_PB)
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
