% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIndex.R
\name{loadIndex}
\alias{loadIndex}
\title{loadIndex}
\usage{
loadIndex(indexPath)
}
\arguments{
\item{indexPath}{The path to the *_tagMeppRindex.fa.gz made by
\code{\link{makeIndex}}.}
}
\value{
A tagMepprIndex-object with the following slots:

\describe{
\item{index}{The path to the .fa.gz of hybrid reference genome}
\item{ITR}{The name of the used transposon-system or ITR.fa}
\item{TIS}{A GRanges-object of all target insertion sites.}
\item{targetInsertionSite}{The targetInsertionSite}
\item{insertName}{The name of the transposon in the hybrid reference}
}
}
\description{
Load a hybrid reference genome
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

# stuff happens and reference_T42 is no longer loaded, so lets load it:
reference_hg19_PB = loadIndex(indexPath = "/home/A.Dent/analysis42/BSgenome.Hsapiens.UCSC.hg19_PiggyBac_tagMeppRindex.fa.gz")
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
