% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIndex.R
\name{makeIndex}
\alias{makeIndex}
\title{makeIndex}
\usage{
makeIndex(indexPath, bsgenome = NULL, ITR = "PiggyBac",
  targetInsertionSite = "TTAA", blockSizeMult = NULL, verbose = F)
}
\arguments{
\item{indexPath}{A path to the folder to store the reference-files.}

\item{bsgenome}{A BSgenome onject (default: BSgenome.Hsapiens.UCSC.hg19)}

\item{ITR}{Can take PiggyBac (default), SleepingBeauty, or a path to a 1000xN-padded ITR.fasta.}

\item{targetInsertionSite}{The target insertion site of your system ("TTAA" for piggyBac, "TA" for sleepingBeauty)}

\item{blockSizeMult}{Set a blocksize multiplication factor for BWA index. If
left NULL, will use calculate optimal value. Change only when absolutely
needed and when you have read
\href{https://github.com/lh3/bwa/issues/104}{this thread}.}

\item{verbose}{Do you want a more chatty function?}
}
\value{
Resulting files stored as [indexPath]/[bsgenome]_[ITR]_tagMeppRindex.*.
The .fa file will be the hybrid reference genome, while the .tis file stores
all possible insertions sites.

The function also returns a tagMepprIndex-object with the following slots:

\describe{
\item{index}{The path to the hybrid reference genome}
\item{ITR}{The name of the used transposon-system or ITR.fa}
\item{TIS}{A GRanges-object of all target insertion sites.}
\item{targetInsertionSite}{The targetInsertionSite}
\item{insertName}{The name of the transposon in the hybrid reference}
}
}
\description{
Generate a hybrid reference genome
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
