% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTagMeppr.R
\name{newTagMeppr}
\alias{newTagMeppr}
\title{newTagMeppr}
\usage{
newTagMeppr(F1, F2, R1, R2, name, protocol = "PiggyBac")
}
\arguments{
\item{F1}{The path to forward_R1.fastq.gz.}

\item{F2}{The path to forward_R2.fastq.gz.}

\item{R1}{The path to reverse_R1.fastq.gz.}

\item{R2}{The path to reverse_R2.fastq.gz.}

\item{name}{The name of the sample.}

\item{protocol}{Either PiggyBac or SleepingBeauty are supported, but other
protocols are allowed.}
}
\value{
A tagMepprSample-object with the following slots:

\describe{
\item{F1}{The path of forward_R1.fastq.gz.}
\item{F2}{The path of forward_R2.fastq.gz.}
\item{R1}{The path of reverse_R1.fastq.gz.}
\item{R2}{The path of reverse_R2.fastq.gz.}
\item{name}{The name of the sample.}
\item{protocol}{The name of the protocol.}
}
}
\description{
Construct a tagMepprExperiment-object from a tagmap-sample.
}
\examples{
\dontrun{
C1 = newTagMeppr(F1 = '/home/A.Dent/analysis42/clone1_FWD_R1.fq.gz',
                 F2 = '/home/A.Dent/analysis42/clone1_FWD_R2.fq.gz',
                 R1 = '/home/A.Dent/analysis42/clone1_REV_R1.fq.gz',
                 R2 = '/home/A.Dent/analysis42/clone1_REC_R2.fq.gz',
                 name = "clone1",
                 protocol = 'PiggyBac')

}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
