% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PiggyBacITRs}
\alias{PiggyBacITRs}
\title{PiggyBacITRs}
\format{An object of class \code{DNAStringSet} of length 1.}
\source{
\url{https://blog.addgene.org/piggybac-ing-through-the-genome-editing-field}
}
\usage{
PiggyBacITRs
}
\description{
A DNAStringSet instance of the inverted terminal repeat (ITR) sequences of
the Piggybac transposon. The sequences are padded with a "cargo" of 1000 Ns.
}
\keyword{datasets}
