% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInsertions.R
\name{plotInsertions}
\alias{plotInsertions}
\title{plotInsertions}
\usage{
plotInsertions(exp, chrom = paste0("chr", c(1:25, "X")),
  colours = NULL, sideBySide = F)
}
\arguments{
\item{exp}{A list or singular instance of tagMeppr-object(s):
first run \code{\link{findInsertions}}.}

\item{chrom}{Which chromosomes should be plotted (even when no insertions are found).}

\item{colours}{Choose colours for individual tagMeppr-objects (if sideBySide = T).}

\item{sideBySide}{If true, will show per-object ideogram. If false, will show
all (colour-coded) objects inside one ideogram.}
}
\value{
An object of the class ggplot.
}
\description{
Identify highly likely insertion-sites in an tagMeppr-object.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C1 = newTagMeppr(F1 = '/home/A.Dent/analysis42/clone1_FWD_R1.fq.gz',
                 F2 = '/home/A.Dent/analysis42/clone1_FWD_R2.fq.gz',
                 R1 = '/home/A.Dent/analysis42/clone1_REV_R1.fq.gz',
                 R2 = '/home/A.Dent/analysis42/clone1_REC_R2.fq.gz',
                 name = "clone1",
                 protocol = 'PiggyBac')

align(exp = C1, ref = reference_hg19_PB, cores = 30, empericalCentre = T)

findInsertions(exp = C1, ref = reference_hg19_PB, minP = 0.05)

plotInsertions(exp = C1, chrom = c('chr1','chr2'))

}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
