% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSite}
\alias{plotSite}
\title{plotSite}
\usage{
plotSite(exp, site = 1, forceDetail = F, maxReads = Inf)
}
\arguments{
\item{exp}{A list or singular instance of tagMeppr-object(s):
first run \code{\link{findInsertions}}.}

\item{site}{Which insertion should be plotted?}

\item{forceDetail}{Even with more than 1000reds, still use the high-detail version.}

\item{maxReads}{Set the maximum number of reads. If forward and/or reverse
exceeds this, they will be downsampled to this number.}
}
\value{
An object of the class ggplot.
}
\description{
Plot the read distribution of an insertion.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C1 = newTagMeppr(F1 = '/home/A.Dent/analysis42/clone1_FWD_R1.fq.gz',
                 F2 = '/home/A.Dent/analysis42/clone1_FWD_R2.fq.gz',
                 R1 = '/home/A.Dent/analysis42/clone1_REV_R1.fq.gz',
                 R2 = '/home/A.Dent/analysis42/clone1_REC_R2.fq.gz',
                 name = "clone1",
                 protocol = 'PiggyBac')

align(exp = C1, ref = reference_hg19_PB, cores = 30, empericalCentre = T)

findInsertions(exp = C1, ref = reference_hg19_PB, minP = 0.05)

plotSite(exp = C1, site = 1)

}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
