% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tagMepprSample}
\alias{print.tagMepprSample}
\title{print}
\usage{
\method{print}{tagMepprSample}(x, ...)
}
\arguments{
\item{x}{A tagMepprSample object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{NULL}
}
\description{
print
}
