% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{results}
\alias{results}
\title{results}
\usage{
results(tagMepprSample, alpha = 1, countThreshold = 0)
}
\arguments{
\item{tagMepprSample}{Your sample}

\item{alpha}{A p-value treshold (default: 1)}

\item{countThreshold}{A p-value treshold (default: 0)}

\item{orientation}{Select insertions with a specific orientation}

\item{...}{other arguments}
}
\description{
Obtain the results
}
\details{
Running this will generate a data.frame of results
}
\examples{

results(tagMepprSample, alpha = 0.05, countThreshold = 100)

}
