% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runIDgen}
\alias{runIDgen}
\title{runIDgen}
\usage{
runIDgen(n = 1)
}
\arguments{
\item{n}{The number of ID's needed.}
}
\description{
Generate an unique 10-item ID
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
