% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SleepingBeautyITRs}
\alias{SleepingBeautyITRs}
\title{SleepingBeautyITRs}
\format{An object of class \code{DNAStringSet} of length 1.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/9390559}
}
\usage{
SleepingBeautyITRs
}
\description{
A DNAStringSet instance of the inverted terminal repeat (ITR) sequences of
the SleepingBeauty transposon. The sequences are padded with a "cargo" of 1000 Ns.
}
\keyword{datasets}
