% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tagMepprCol}
\alias{tagMepprCol}
\title{tagMepprCol}
\usage{
tagMepprCol(n)
}
\arguments{
\item{n}{The number of colours needed.}
}
\description{
A pallette based on CYRUP (github.com/robinweide/CYRUP)
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
