% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{align}
\usage{
align(exp, ref, cores = 20, dedup = T, empericalCentre = F,
  verbose = F)
}
\arguments{
\item{exp}{The tagMeppr-object of a sample.}

\item{ref}{A reference object of \code{\link{makeIndex}} or \code{\link{loadIndex}}.}

\item{cores}{The number of threads used in the mapping-stage.}

\item{dedup}{Remove suspected PCR-duplicates?}

\item{empericalCentre}{The location of the center of the inserted sequence
in the transposon.fa. This is needed to find the orientation of the
insertion. Can be T, F or numeric:
\describe{
  \item{TRUE}{Will find it automatically}
  \item{FALSE}{Will take the middle of the transposon.fa. This doesn't
have to be the case when you loaded a BAC containing the insertion.}
  \item{numeric}{Will use this number as the postion of the middle.}
}}

\item{verbose}{Do you want a more chatty function?}
}
\value{
The object will be updated with the following slots:
\describe{
\item{alignmentFolder}{The path to the folder of BAM-files}
\item{insertName}{The name of the transposon in the hybrid reference}
\item{seqinfo}{The seqinfo of the hybrid reference}
\item{FWDBAM}{GenomicAlignments of the mapping of forward reads}
\item{REVBAM}{GenomicAlignments of the mapping of reverse reads}
\item{alignedReadsFWD}{Granges-oject with informative forward reads}
\item{alignedReadsREV}{Granges-oject with informative reverse reads}
\item{insertionMid}{The used position of the middle of the insertion}
}
}
\description{
Align reads on the hubrid reference genome.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C9 = newTagMeppr(F1 = 'clone9_FWD_R1.fq.gz',
                 F2 = 'clone9_FWD_R2.fq.gz',
                 R1 = 'clone9_REV_R1.fq.gz',
                 R2 = 'clone9_REV_R2.fq.gz',
                 name = "clone9",
                 protocol = 'PiggyBac')

align(exp = C9, ref = reference_hg19_PB, cores = 30, empericalCentre = T)
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
