% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInsertions.R
\name{findInsertions}
\alias{findInsertions}
\title{findInsertions}
\usage{
findInsertions(exp, ref, minP = 0.05)
}
\arguments{
\item{exp}{The tagMeppr-object of a sample: first run \code{\link{align}}.}

\item{ref}{A reference object of \code{\link{makeIndex}} or \code{\link{loadIndex}}.}

\item{minP}{The minimal adjusted p-value.}
}
\value{
The object will be updated with a results-table:
\describe{
\item{cnt_FWD}{Forward read-counts at this insertion-site}
\item{orientation_FWD}{Forward read-orientation at the insertion}
\item{D_FWD}{The imbalance of forward reads on both sides of the
site: D = 0 means no imbalance}
\item{pBinom_FWD}{The probability of the imbalance}
\item{cnt_REV}{Reverse read-counts at this insertion-site}
\item{orientation_REV}{Reverse read-orientation at the insertion}
\item{D_REV}{The imbalance of forward reads on both sides of the site: D = 0
means no imbalance}
\item{pBinom_REV}{The probability of the imbalance}
\item{pMarkerMeppR}{The combined probability of the imbalances}
\item{padj}{Bonferroni-adjusted probabilties}
\item{orientation}{The orientation of the insertion}
}
}
\description{
Identify highly likely insertion-sites in an tagMeppr-object.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C1 = newTagMeppr(F1 = '/home/A.Dent/analysis42/clone1_FWD_R1.fq.gz',
                 F2 = '/home/A.Dent/analysis42/clone1_FWD_R2.fq.gz',
                 R1 = '/home/A.Dent/analysis42/clone1_REV_R1.fq.gz',
                 R2 = '/home/A.Dent/analysis42/clone1_REC_R2.fq.gz',
                 name = "clone1",
                 protocol = 'PiggyBac')

align(exp = C1, ref = reference_hg19_PB, cores = 30, empericalCentre = T)

findInsertions(exp = C1, ref = reference_T42, minP = 0.05)
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
