% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInsertions.R
\name{plotInsertions}
\alias{plotInsertions}
\title{plotInsertions}
\usage{
plotInsertions(exp, chrom = paste0("chr", c(1:25, "X")),
  colours = NULL, sideBySide = F, p = 0.05, showOrientation = F)
}
\arguments{
\item{exp}{A list or singular instance of tagMeppr-object(s):
first run \code{\link{findInsertions}}.}

\item{chrom}{Which chromosomes should be plotted (even when no insertions are found).}

\item{colours}{Choose colours for individual tagMeppr-objects (if sideBySide = T).}

\item{sideBySide}{If true, will show per-object ideogram. If false, will show}

\item{p}{Set a p-value treshold
all (colour-coded) objects inside one ideogram.}

\item{showOrientation}{Split hits on orientation.}
}
\value{
An object of the class ggplot.
}
\description{
Identify highly likely insertion-sites in an tagMeppr-object.
}
\examples{
\dontrun{
reference_hg19_PB = makeIndex(indexPath = '/home/A.Dent/analysis42/',
                              bsgenome = 'BSgenome.Hsapiens.UCSC.hg19',
                              ITR = 'PiggyBac')

C9 = newTagMeppr(F1 = 'clone9_FWD_R1.fq.gz',
                 F2 = 'clone9_FWD_R2.fq.gz',
                 R1 = 'clone9_REV_R1.fq.gz',
                 R2 = 'clone9_REV_R2.fq.gz',
                 name = "clone9",
                 protocol = 'PiggyBac')

align(exp = C9, ref = reference_hg19_PB, cores = 30, empericalCentre = T)

findInsertions(exp = C9, ref = reference_hg19_PB)

plotInsertions(exp = C9, chrom = c('chr1','chr2'))

}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
